Move getDropDownOptionsWithFavouriteDashboardItems to utils
